+++ src/components/templates/TasksDashboardStudent.unit.js
@@ -1,12 +1,10 @@
 import TasksDashboardStudent from "./TasksDashboardStudent";
 import TasksList from "@components/organisms/TasksList";
 import vCustomEmptyState from "@components/molecules/vCustomEmptyState";
-import {
-	overDueTasks,
-	openTasksWithoutDueDate,
-	openTasksWithDueDate,
-} from "@@/stories/mockData/Tasks";
+import mocks from "@@/stories/mockData/Tasks";
 import Vuetify from "vuetify";
+
+const { overDueTasks, openTasksWithoutDueDate, openTasksWithDueDate } = mocks;
 
 describe("@components/templates/TasksDashboardStudent", () => {
 	const mockStore = {
@@ -18,10 +16,10 @@ describe("@components/templates/TasksDashboardStudent", () => {
 					noDueDate: openTasksWithoutDueDate,
 				}),
 				getStatus: () => "completed",
-				hasNoTasks: () => false,
+				hasTasks: () => true,
 				getCompletedTasksForStudent: () => ({ submitted: [], graded: [] }),
-				hasNoOpenTasksStudent: () => false,
-				hasNoCompletedTasks: () => true,
+				hasOpenTasksStudent: () => true,
+				hasCompletedTasks: () => false,
 				hasFilterSelected: () => false,
 			},
 		},